requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

Babel>=0.9.6

Django>=1.4.2,<1.7
django_openstack_auth>=1.1.7,!=1.1.8
python-keystoneclient-kerberos
